Module rustfmt_nightly::attr

Format attributes and meta items.

Modules

doc_comment ğŸ”’

Traits

MetaVisitor ğŸ”’

Functions

argument_shape ğŸ”’

attr_prefix ğŸ”’

contains_name ğŸ”’

filter_inline_attrs ğŸ”’

Returns attributes that are within outer_span.

first_attr_value_str_by_name ğŸ”’

format_derive ğŸ”’

get_attrs_from_stmt ğŸ”’

Returns attributes on the given statement.

get_span_without_attrs ğŸ”’

has_newlines_before_after_comment ğŸ”’

is_derive ğŸ”’

rewrite_initial_doc_comments ğŸ”’

Rewrite the any doc comments which come before any other attributes.

take_while_with_pred ğŸ”’

Returns the first group of attributes that fills the given predicate. We consider two doc comments are in different group if they are separated by normal comments.