Enum rustfmt_nightly::comment::FullCodeCharKind

pub(crate) enum FullCodeCharKind {
    Normal,
    StartComment,
    InComment,
    EndComment,
    StartStringCommented,
    EndStringCommented,
    InStringCommented,
    StartString,
    EndString,
    InString,
}

Distinguish between functional part of code and comments, describing opening and closing of comments for ease when chunking code from tagged characters

Variants

Normal

StartComment

The first character of a comment, there is only one for a comment (always ‘/’)

InComment

Any character inside a comment including the second character of comment marks (“//”, “/*”)

EndComment

Last character of a comment, ‘\n’ for a line comment, ‘/’ for a block comment.

StartStringCommented

Start of a mutlitine string inside a comment

EndStringCommented

End of a mutlitine string inside a comment

InStringCommented

Inside a commented string

StartString

Start of a mutlitine string

EndString

End of a mutlitine string

InString

Inside a string.

Implementations

impl FullCodeCharKind

pub(crate) fn is_comment(self) -> bool

pub(crate) fn inside_comment(self) -> bool

Returns true if the character is inside a comment

pub(crate) fn is_string(self) -> bool

pub(crate) fn is_commented_string(self) -> bool

Returns true if the character is within a commented string

fn to_codecharkind(self) -> CodeCharKind