Struct rustfmt_nightly::comment::CommentRewrite

struct CommentRewrite<'a> {

    result: String,
    code_block_buffer: String,
    is_prev_line_multi_line: bool,
    code_block_attr: Option<CodeBlockAttribute>,
    item_block: Option<ItemizedBlock>,
    comment_line_separator: String,
    indent_str: String,
    max_width: usize,
    fmt_indent: Indent,
    fmt: StringFormat<'a>,
    opener: String,
    closer: String,
    line_start: String,
    style: CommentStyle<'a>,

}

Fields
result: Stringcode_block_buffer: Stringis_prev_line_multi_line: boolcode_block_attr: Option<CodeBlockAttribute>item_block: Option<ItemizedBlock>comment_line_separator: Stringindent_str: Stringmax_width: usizefmt_indent: Indentfmt: StringFormat<'a>opener: Stringcloser: Stringline_start: Stringstyle: CommentStyle<'a>
Implementations

impl<'a> CommentRewrite<'a>

fn new( orig: &'a str, block_style: bool, shape: Shape, config: &'a Config ) -> CommentRewrite<'a>

fn join_block(s: &str, sep: &str) -> String

fn buffer_contains_comment(&self) -> bool

Check if any characters were written to the result buffer after the start of the comment. when calling CommentRewrite::new() the result buffer is initiazlied with the opening characters for the comment.

fn finish(self) -> String

fn handle_line( &mut self, orig: &'a str, i: usize, line: &'a str, has_leading_whitespace: bool, is_doc_comment: bool ) -> bool