rustfmt_nightly::comment
Struct ItemizedBlock

struct ItemizedBlock {
    lines: Vec<String>,
    indent: usize,
    opener: String,
    line_start: String,
}

Block that is formatted as an item.

An item starts with either a star *, a dash -, a greater-than >, a plus ‘+’, or a number 12. or 34) (with at most 2 digits). An item represents CommonMark’s “list items” and/or “block quotes”, but note that only a subset of CommonMark is recognized - see the doc comment of ItemizedBlock::get_marker_length for more details.

Different level of indentation are handled by shrinking the shape accordingly.

Fields
lines: Vec<String>

the lines that are identified as part of an itemized block

indent: usize

the number of characters (typically whitespaces) up to the item marker

opener: String

the string that marks the start of an item

line_start: String

sequence of characters (typically whitespaces) to prefix new lines that are part of the item

Implementations

impl ItemizedBlock

fn get_marker_length(trimmed: &str) -> Option<usize>

Checks whether the trimmed line includes an item marker. Returns None if there is no marker. Returns the length of the marker (in bytes) if one is present. Note that the length includes the whitespace that follows the marker, for example the marker in "* list item" has the length of 2.

This function recognizes item markers that correspond to CommonMark’s “bullet list marker”, “block quote marker”, and/or “ordered list marker”.

Compared to CommonMark specification, the number of digits that are allowed in an “ordered list marker” is more limited (to at most 2 digits). Limiting the length of the marker helps reduce the risk of recognizing arbitrary numbers as markers. See also https://talk.commonmark.org/t/blank-lines-before-lists-revisited/1990 which gives the following example where a number (i.e. “1868”) doesn’t signify an ordered list:

The Captain died in
1868. He wes buried in...

fn new(line: &str) -> Option<ItemizedBlock>

Creates a new ItemizedBlock described with the given line. Returns None if line doesn’t start an item.

fn create_string_format<'a>( &'a self, fmt: &'a StringFormat<'_>, ) -> StringFormat<'a>

Returns a StringFormat used for formatting the content of an item.

fn add_line(&mut self, line: &str) -> bool

Returns true if the line is part of the current itemized block. If it is, then it is added to the internal lines list.

fn trimmed_block_as_string(&self) -> String

Returns the block as a string, with each line trimmed at the start.

fn original_block_as_string(&self) -> String

Returns the block as a string under its original form.