rustfmt_nightly::formatting
Enum Timer

enum Timer {
    Disabled,
    Initialized(Instant),
    DoneParsing(Instant, Instant),
    DoneFormatting(Instant, Instant, Instant),
}

Variants

Disabled

Initialized(Instant)

DoneParsing(Instant, Instant)

DoneFormatting(Instant, Instant, Instant)

Implementations

impl Timer

fn start() -> Timer

fn done_parsing(self) -> Self

fn done_formatting(self) -> Self

fn get_parse_time(&self) -> f32

Returns the time it took to parse the source files in seconds.

fn get_format_time(&self) -> f32

Returns the time it took to go from the parsed AST to the formatted output. Parsing time is not included.

fn duration_to_f32(d: Duration) -> f32