Module rustfmt_nightly::formatting::newline_style

Enums

EffectiveNewlineStyle ğŸ”’

Constants

CARRIAGE_RETURN ğŸ”’

LINE_FEED ğŸ”’

UNIX_NEWLINE ğŸ”’

WINDOWS_NEWLINE ğŸ”’

Functions

apply_newline_style ğŸ”’

Apply this newline style to the formatted text. When the style is set to Auto, the raw_input_text is used to detect the existing line endings.

auto_detect_newline_style ğŸ”’

convert_to_unix_newlines ğŸ”’

convert_to_windows_newlines ğŸ”’

effective_newline_style ğŸ”’

native_newline_style ğŸ”’