Struct rustfmt_nightly::macros::MacroArgParser

struct MacroArgParser {
    buf: String,
    start_tok: Token,
    is_meta_var: bool,
    last_tok: Token,
    result: Vec<ParsedMacroArg>,
}

Parses macro arguments on macro def.

Fields
buf: String

Either a name of the next metavariable, a separator, or junk.

start_tok: Token

The first token of the current buffer.

is_meta_var: bool

true if we are parsing a metavariable or a repeat.

last_tok: Token

The last token parsed.

result: Vec<ParsedMacroArg>

Holds the parsed arguments.

Implementations

impl MacroArgParser

fn new() -> MacroArgParser

fn set_last_tok(&mut self, tok: &TokenTree)

fn add_separator(&mut self)

fn add_other(&mut self)

fn add_meta_variable(&mut self, iter: &mut Cursor) -> Option<()>

fn add_delimited(&mut self, inner: Vec<ParsedMacroArg>, delim: Delimiter)

fn add_repeat(
&mut self,
inner: Vec<ParsedMacroArg>,
delim: Delimiter,
iter: &mut Cursor
) -> Option<()>

fn update_buffer(&mut self, t: &Token)

fn need_space_prefix(&self) -> bool

fn parse(self, tokens: TokenStream) -> Option<Vec<ParsedMacroArg>>

Returns a collection of parsed macro def’s arguments.