Struct rustfmt_nightly::modules::ModResolver

pub(crate) struct ModResolver<'ast, 'sess> {
    parse_sess: &'sess ParseSess,
    directory: Directory,
    file_map: BTreeMap<FileName, Module<'ast>>,
    recursive: bool,
}

Maps each module to the corresponding file.

Fields
parse_sess: &'sess ParseSessdirectory: Directoryfile_map: BTreeMap<FileName, Module<'ast>>recursive: bool
Implementations

impl<'ast, 'sess, 'c> ModResolver<'ast, 'sess>

pub(crate) fn new( parse_sess: &'sess ParseSess, directory_ownership: DirOwnership, recursive: bool ) -> Self

Creates a new ModResolver.

pub(crate) fn visit_crate( self, krate: &'ast Crate ) -> Result<BTreeMap<FileName, Module<'ast>>, ModuleResolutionError>

Creates a map that maps a file name to the module in AST.

fn visit_cfg_if( &mut self, item: Cow<'ast, Item> ) -> Result<(), ModuleResolutionError>

Visit cfg_if macro and look for module declarations.

fn visit_mod_outside_ast( &mut self, items: ThinVec<P<Item>> ) -> Result<(), ModuleResolutionError>

Visit modules defined inside macro calls.

fn visit_mod_from_ast( &mut self, items: &'ast [P<Item>] ) -> Result<(), ModuleResolutionError>

Visit modules from AST.

fn visit_sub_mod( &mut self, item: &'c Item, sub_mod: Module<'ast> ) -> Result<(), ModuleResolutionError>

fn peek_sub_mod( &self, item: &'c Item, sub_mod: &Module<'ast> ) -> Result<Option<SubModKind<'c, 'ast>>, ModuleResolutionError>

Inspect the given sub-module which we are about to visit and returns its kind.

fn insert_sub_mod( &mut self, sub_mod_kind: SubModKind<'c, 'ast> ) -> Result<(), ModuleResolutionError>

fn visit_sub_mod_inner( &mut self, sub_mod: Module<'ast>, sub_mod_kind: SubModKind<'c, 'ast> ) -> Result<(), ModuleResolutionError>

fn visit_sub_mod_after_directory_update( &mut self, sub_mod: Module<'ast>, directory: Option<Directory> ) -> Result<(), ModuleResolutionError>

fn find_external_module( &self, mod_name: Ident, attrs: &[Attribute], sub_mod: &Module<'ast> ) -> Result<Option<SubModKind<'c, 'ast>>, ModuleResolutionError>

Find a file path in the filesystem which corresponds to the given module.

fn push_inline_mod_directory(&mut self, id: Ident, attrs: &[Attribute])

fn find_mods_outside_of_ast( &self, attrs: &[Attribute], sub_mod: &Module<'ast> ) -> Vec<(PathBuf, DirOwnership, Module<'ast>)>