Module rustfmt_nightly::overflow

Rewrite a list some items with overflow.

Macros

impl_into_overflowable_item_for_ast_node ğŸ”’

impl_into_overflowable_item_for_rustfmt_types ğŸ”’

Structs

Context ğŸ”’

Enums

OverflowableItem ğŸ”’

Constants

SPECIAL_CASE_ATTR ğŸ”’

SPECIAL_CASE_MACROS ğŸ”’

A list of format!-like macros, that take a long format string and a list of arguments to format.

Traits

IntoOverflowableItem ğŸ”’

Functions

can_be_overflowed ğŸ”’

into_overflowable_list ğŸ”’

last_item_shape ğŸ”’

Returns a shape for the last argument which is going to be overflowed.

maybe_get_args_offset ğŸ”’

In case special-case style is required, returns an offset from which we start horizontal layout.

need_block_indent ğŸ”’

no_long_items ğŸ”’

rewrite_with_angle_brackets ğŸ”’

rewrite_with_parens ğŸ”’

rewrite_with_square_brackets ğŸ”’

shape_from_indent_style ğŸ”’