Module rustfmt_nightly::string

Structs

StringFormat ğŸ”’

Describes the layout of a piece of text.

Enums

SnippetState ğŸ”’

Result of breaking a string so it fits in a line and the state it ended in. The state informs about what to do with the snippet and how to continue the breaking process.

Constants

MIN_STRING ğŸ”’

Functions

break_string ğŸ”’

Break the input string at a boundary character around the offset max_width. A boundary character is either a punctuation or a whitespace. FIXME(issue#3281): We must follow UAX#14 algorithm instead of this.

detect_url ğŸ”’

Returns the index to the end of the URL if the split at index of the given string includes a URL or alike. Otherwise, returns None.

graphemes_width ğŸ”’

is_new_line ğŸ”’

is_part_of_type ğŸ”’

is_punctuation ğŸ”’

is_valid_linebreak ğŸ”’

is_whitespace ğŸ”’

not_whitespace_except_line_feed ğŸ”’

rewrite_string ğŸ”’

trim_end_but_line_feed ğŸ”’

Trims whitespaces to the right except for the line feed character.