Struct rustfmt_nightly::FileLines

pub struct FileLines(Option<HashMap<FileName, Vec<Range>>>);

A set of lines in files.

It is represented as a multimap keyed on file names, with values a collection of non-overlapping ranges sorted by their start point. An inner None is interpreted to mean all lines in all files.

Tuple Fields
0: Option<HashMap<FileName, Vec<Range>>>
Implementations

impl FileLines

pub(crate) fn all() -> FileLines

Creates a FileLines that contains all lines in all files.

pub(crate) fn is_all(&self) -> bool

Returns true if this FileLines contains all lines in all files.

pub fn from_ranges(ranges: HashMap<FileName, Vec<Range>>) -> FileLines

pub fn files(&self) -> Files<'_>

Returns an iterator over the files contained in self.

pub fn to_json_spans(&self) -> Vec<JsonSpan>

Returns JSON representation as accepted by the --file-lines JSON arg.

fn file_range_matches<F>(&self, file_name: &FileName, f: F) -> boolwhere
F: FnMut(&Range) -> bool,

Returns true if self includes all lines in all files. Otherwise runs f on all ranges in the designated file (if any) and returns true if f ever does.

pub(crate) fn contains(&self, range: &LineRange) -> bool

Returns true if range is fully contained in self.

pub(crate) fn intersects(&self, range: &LineRange) -> bool

Returns true if any lines in range are in self.

pub(crate) fn contains_line(&self, file_name: &FileName, line: usize) -> bool

Returns true if line from file_name is in self.

pub(crate) fn contains_range(
&self,
file_name: &FileName,
lo: usize,
hi: usize
) -> bool

Returns true if all the lines between lo and hi from file_name are in self.