Struct rustfmt_nightly::Session

pub struct Session<'b, T: Write> {
    pub config: Config,
    pub out: Option<&'b mut T>,
    pub(crate) errors: ReportedErrors,
    pub(crate) source_file: Vec<(FileName, String)>,
    pub(crate) emitter: Box<dyn Emitter + 'b>,
}

A session is a run of rustfmt across a single or multiple inputs.

Fields
config: Configout: Option<&'b mut T>errors: ReportedErrorssource_file: Vec<(FileName, String)>emitter: Box<dyn Emitter + 'b>
Implementations

impl<'b, T: Write + 'b> Session<'b, T>

pub(crate) fn format_input_inner( &mut self, input: Input, is_macro_def: bool ) -> Result<FormatReport, ErrorKind>

impl<'b, T: Write + 'b> Session<'b, T>

pub fn new(config: Config, out: Option<&'b mut T>) -> Session<'b, T>

pub fn format(&mut self, input: Input) -> Result<FormatReport, ErrorKind>

The main entry point for Rustfmt. Formats the given input according to the given config. out is only necessary if required by the configuration.

pub fn override_config<F, U>(&mut self, config: Config, f: F) -> Uwhere F: FnOnce(&mut Session<'b, T>) -> U,

pub fn add_operational_error(&mut self)

pub fn has_operational_errors(&self) -> bool

pub fn has_parsing_errors(&self) -> bool

pub fn has_formatting_errors(&self) -> bool

pub fn has_check_errors(&self) -> bool

pub fn has_diff(&self) -> bool

pub fn has_unformatted_code_errors(&self) -> bool

pub fn has_no_errors(&self) -> bool