Module rustfmt_nightly::vertical

Traits

AlignedItem ğŸ”’

Functions

group_aligned_items ğŸ”’

Returns the index in fields up to which a field belongs to the current group. The returned string is the group separator to use when rewriting the fields. Groups are defined by blank lines.

rewrite_aligned_items_inner ğŸ”’

rewrite_with_alignment ğŸ”’

struct_field_prefix_max_min_width ğŸ”’