Struct rustfmt_nightly::visitor::SnippetProvider

pub(crate) struct SnippetProvider {
    big_snippet: Lrc<String>,
    start_pos: usize,
    end_pos: usize,
}

Creates a string slice corresponding to the specified span.

Fields
big_snippet: Lrc<String>

A pointer to the content of the file we are formatting.

start_pos: usize

A position of the start of big_snippet, used as an offset.

end_pos: usize

An end position of the file that this snippet lives.

Implementations

impl SnippetProvider

pub(crate) fn span_to_snippet(&self, span: Span) -> Option<&str>

pub(crate) fn new(
start_pos: BytePos,
end_pos: BytePos,
big_snippet: Lrc<String>
) -> Self

pub(crate) fn entire_snippet(&self) -> &str

pub(crate) fn start_pos(&self) -> BytePos

pub(crate) fn end_pos(&self) -> BytePos