Module rustfmt_nightly::config

Re-exports

pub use crate::config::file_lines::FileLines;

pub use crate::config::file_lines::FileName;

pub use crate::config::file_lines::Range;

pub use crate::config::macro_names::MacroSelector;

pub use crate::config::macro_names::MacroSelectors;

pub use crate::config::lists::*;

pub use crate::config::options::*;

Modules

config_type 🔒

file_lines 🔒

This module contains types and functions to support formatting specific line ranges.

lists 🔒

Configuration options related to rewriting a list.

macro_names 🔒

This module contains types and functions to support formatting specific macros.

options 🔒

Structs

Config

ConfigSetter

ConfigWasSet

PartialConfig

ToTomlError

Functions

config_path 🔒

get_toml_path 🔒

load_config

Loads a config by checking the client-supplied options and if appropriate, the file system (including searching the file system for overrides).