Module rustfmt_nightly::config::file_lines

This module contains types and functions to support formatting specific line ranges.

Structs

FileLines

A set of lines in files.

Files

FileLines files iterator.

JsonSpan

LineRange

A range of lines in a file, inclusive of both ends.

Range

A range that is inclusive of both ends.

Enums

FileLinesError

FileName

Defines the name of an input - either a file or stdin.

Functions

canonicalize_path_string 🔒

normalize_ranges 🔒

Normalizes the ranges so that the invariants for FileLines hold: ranges are non-overlapping, and ordered by their start point.