rustfmt_nightly::config
Module options

Re-exports

pub use __define_config_type_on_enum_NewlineStyle::NewlineStyle;

pub use __define_config_type_on_enum_BraceStyle::BraceStyle;

pub use __define_config_type_on_enum_ControlBraceStyle::ControlBraceStyle;

pub use __define_config_type_on_enum_IndentStyle::IndentStyle;

pub use __define_config_type_on_enum_Density::Density;

pub use __define_config_type_on_enum_TypeDensity::TypeDensity;

pub use __define_config_type_on_enum_Heuristics::Heuristics;

pub use __define_config_type_on_enum_GroupImportsTactic::GroupImportsTactic;

pub use __define_config_type_on_enum_ImportGranularity::ImportGranularity;

pub use __define_config_type_on_enum_HexLiteralCase::HexLiteralCase;

pub use __define_config_type_on_enum_ReportTactic::ReportTactic;

pub use __define_config_type_on_enum_EmitMode::EmitMode;

pub use __define_config_type_on_enum_Color::Color;

pub use __define_config_type_on_enum_Version::Version;

pub use __define_config_type_on_enum_Verbosity::Verbosity;

pub use __define_config_type_on_enum_Edition::Edition;

pub use __define_config_type_on_enum_MatchArmLeadingPipe::MatchArmLeadingPipe;

pub use __define_config_type_on_enum_StyleEdition::StyleEdition;

Modules

__define_config_type_on_enum_BraceStyle 🔒

__define_config_type_on_enum_Color 🔒

__define_config_type_on_enum_ControlBraceStyle 🔒

__define_config_type_on_enum_Density 🔒

__define_config_type_on_enum_Edition 🔒

__define_config_type_on_enum_EmitMode 🔒

__define_config_type_on_enum_GroupImportsTactic 🔒

__define_config_type_on_enum_Heuristics 🔒

__define_config_type_on_enum_HexLiteralCase 🔒

__define_config_type_on_enum_ImportGranularity 🔒

__define_config_type_on_enum_IndentStyle 🔒

__define_config_type_on_enum_MatchArmLeadingPipe 🔒

__define_config_type_on_enum_NewlineStyle 🔒

__define_config_type_on_enum_ReportTactic 🔒

__define_config_type_on_enum_StyleEdition 🔒

__define_config_type_on_enum_TypeDensity 🔒

__define_config_type_on_enum_Verbosity 🔒

__define_config_type_on_enum_Version 🔒

Structs

ArrayWidth

AttrFnLikeWidth

BinopSeparator

BlankLinesLowerBound

BlankLinesUpperBound

BraceStyleConfig

ChainWidth

ColorConfig

CombineControlExpr

CommentWidth

CondenseWildcardSuffixes

ControlBraceStyleConfig

DisableAllFormatting

DocCommentCodeBlockWidth

EditionConfig

EmitModeConfig

EmptyItemSingleLine

EnumDiscrimAlignThreshold

ErrorOnLineOverflow

ErrorOnUnformatted

FileLinesConfig

FnArgsLayout

FnCallWidth

FnParamsLayout

FnSingleLine

ForceExplicitAbi

ForceMultilineBlocks

FormatCodeInDocComments

FormatGeneratedFiles

FormatMacroBodies

FormatMacroMatchers

FormatStrings

GeneratedMarkerLineSearchLimit

GroupImportsTacticConfig

HardTabs

HexLiteralCaseConfig

HideParseErrors

Ignore

IgnoreList

A set of directories, files and modules that rustfmt should ignore.

ImportsGranularityConfig

ImportsIndent

ImportsLayout

IndentStyleConfig

InlineAttributeWidth

MakeBackup

MatchArmBlocks

MatchArmLeadingPipeConfig

MatchBlockTrailingComma

MaxWidth

MergeDerives

MergeImports

NewlineStyleConfig

NormalizeComments

NormalizeDocAttributes

OverflowDelimitedExpr

PrintMisformattedFileNames

RemoveNestedParens

ReorderImplItems

ReorderImports

ReorderModules

RequiredVersion

ShortArrayElementWidthThreshold

ShowParseErrors

SingleLineIfElseMaxWidth

SingleLineLetElseMaxWidth

SkipChildren

SkipMacroInvocations

SpaceAfterColon

SpaceBeforeColon

SpacesAroundRanges

StructFieldAlignThreshold

StructLitSingleLine

StructLitWidth

StructVariantWidth

StyleEditionConfig

TabSpaces

TrailingComma

TrailingSemicolon

TypePunctuationDensity

UnstableFeatures

UseFieldInitShorthand

UseSmallHeuristics

UseTryShorthand

Verbose

VersionConfig

WhereSingleLine

WidthHeuristics

WidthHeuristicsConfig

WrapComments

Traits

CliOptions

Maps client-supplied options to Rustfmt’s internals, mostly overriding values in a config with values from the command line.