Struct rustfmt_nightly::config::Config

pub struct Config {

    max_width: (Cell<bool>, bool, usize, bool),
    hard_tabs: (Cell<bool>, bool, bool, bool),
    tab_spaces: (Cell<bool>, bool, usize, bool),
    newline_style: (Cell<bool>, bool, NewlineStyle, bool),
    indent_style: (Cell<bool>, bool, IndentStyle, bool),
    use_small_heuristics: (Cell<bool>, bool, Heuristics, bool),
    width_heuristics: (Cell<bool>, bool, WidthHeuristics, bool),
    fn_call_width: (Cell<bool>, bool, usize, bool),
    attr_fn_like_width: (Cell<bool>, bool, usize, bool),
    struct_lit_width: (Cell<bool>, bool, usize, bool),
    struct_variant_width: (Cell<bool>, bool, usize, bool),
    array_width: (Cell<bool>, bool, usize, bool),
    chain_width: (Cell<bool>, bool, usize, bool),
    single_line_if_else_max_width: (Cell<bool>, bool, usize, bool),
    single_line_let_else_max_width: (Cell<bool>, bool, usize, bool),
    wrap_comments: (Cell<bool>, bool, bool, bool),
    format_code_in_doc_comments: (Cell<bool>, bool, bool, bool),
    doc_comment_code_block_width: (Cell<bool>, bool, usize, bool),
    comment_width: (Cell<bool>, bool, usize, bool),
    normalize_comments: (Cell<bool>, bool, bool, bool),
    normalize_doc_attributes: (Cell<bool>, bool, bool, bool),
    format_strings: (Cell<bool>, bool, bool, bool),
    format_macro_matchers: (Cell<bool>, bool, bool, bool),
    format_macro_bodies: (Cell<bool>, bool, bool, bool),
    skip_macro_invocations: (Cell<bool>, bool, MacroSelectors, bool),
    hex_literal_case: (Cell<bool>, bool, HexLiteralCase, bool),
    empty_item_single_line: (Cell<bool>, bool, bool, bool),
    struct_lit_single_line: (Cell<bool>, bool, bool, bool),
    fn_single_line: (Cell<bool>, bool, bool, bool),
    where_single_line: (Cell<bool>, bool, bool, bool),
    imports_indent: (Cell<bool>, bool, IndentStyle, bool),
    imports_layout: (Cell<bool>, bool, ListTactic, bool),
    imports_granularity: (Cell<bool>, bool, ImportGranularity, bool),
    group_imports: (Cell<bool>, bool, GroupImportsTactic, bool),
    merge_imports: (Cell<bool>, bool, bool, bool),
    reorder_imports: (Cell<bool>, bool, bool, bool),
    reorder_modules: (Cell<bool>, bool, bool, bool),
    reorder_impl_items: (Cell<bool>, bool, bool, bool),
    type_punctuation_density: (Cell<bool>, bool, TypeDensity, bool),
    space_before_colon: (Cell<bool>, bool, bool, bool),
    space_after_colon: (Cell<bool>, bool, bool, bool),
    spaces_around_ranges: (Cell<bool>, bool, bool, bool),
    binop_separator: (Cell<bool>, bool, SeparatorPlace, bool),
    remove_nested_parens: (Cell<bool>, bool, bool, bool),
    combine_control_expr: (Cell<bool>, bool, bool, bool),
    short_array_element_width_threshold: (Cell<bool>, bool, usize, bool),
    overflow_delimited_expr: (Cell<bool>, bool, bool, bool),
    struct_field_align_threshold: (Cell<bool>, bool, usize, bool),
    enum_discrim_align_threshold: (Cell<bool>, bool, usize, bool),
    match_arm_blocks: (Cell<bool>, bool, bool, bool),
    match_arm_leading_pipes: (Cell<bool>, bool, MatchArmLeadingPipe, bool),
    force_multiline_blocks: (Cell<bool>, bool, bool, bool),
    fn_args_layout: (Cell<bool>, bool, Density, bool),
    fn_params_layout: (Cell<bool>, bool, Density, bool),
    brace_style: (Cell<bool>, bool, BraceStyle, bool),
    control_brace_style: (Cell<bool>, bool, ControlBraceStyle, bool),
    trailing_semicolon: (Cell<bool>, bool, bool, bool),
    trailing_comma: (Cell<bool>, bool, SeparatorTactic, bool),
    match_block_trailing_comma: (Cell<bool>, bool, bool, bool),
    blank_lines_upper_bound: (Cell<bool>, bool, usize, bool),
    blank_lines_lower_bound: (Cell<bool>, bool, usize, bool),
    edition: (Cell<bool>, bool, Edition, bool),
    version: (Cell<bool>, bool, Version, bool),
    inline_attribute_width: (Cell<bool>, bool, usize, bool),
    format_generated_files: (Cell<bool>, bool, bool, bool),
    merge_derives: (Cell<bool>, bool, bool, bool),
    use_try_shorthand: (Cell<bool>, bool, bool, bool),
    use_field_init_shorthand: (Cell<bool>, bool, bool, bool),
    force_explicit_abi: (Cell<bool>, bool, bool, bool),
    condense_wildcard_suffixes: (Cell<bool>, bool, bool, bool),
    color: (Cell<bool>, bool, Color, bool),
    required_version: (Cell<bool>, bool, String, bool),
    unstable_features: (Cell<bool>, bool, bool, bool),
    disable_all_formatting: (Cell<bool>, bool, bool, bool),
    skip_children: (Cell<bool>, bool, bool, bool),
    hide_parse_errors: (Cell<bool>, bool, bool, bool),
    error_on_line_overflow: (Cell<bool>, bool, bool, bool),
    error_on_unformatted: (Cell<bool>, bool, bool, bool),
    ignore: (Cell<bool>, bool, IgnoreList, bool),
    verbose: (Cell<bool>, bool, Verbosity, bool),
    file_lines: (Cell<bool>, bool, FileLines, bool),
    emit_mode: (Cell<bool>, bool, EmitMode, bool),
    make_backup: (Cell<bool>, bool, bool, bool),
    print_misformatted_file_names: (Cell<bool>, bool, bool, bool),

}

Fields
max_width: (Cell<bool>, bool, usize, bool)hard_tabs: (Cell<bool>, bool, bool, bool)tab_spaces: (Cell<bool>, bool, usize, bool)newline_style: (Cell<bool>, bool, NewlineStyle, bool)indent_style: (Cell<bool>, bool, IndentStyle, bool)use_small_heuristics: (Cell<bool>, bool, Heuristics, bool)width_heuristics: (Cell<bool>, bool, WidthHeuristics, bool)fn_call_width: (Cell<bool>, bool, usize, bool)attr_fn_like_width: (Cell<bool>, bool, usize, bool)struct_lit_width: (Cell<bool>, bool, usize, bool)struct_variant_width: (Cell<bool>, bool, usize, bool)array_width: (Cell<bool>, bool, usize, bool)chain_width: (Cell<bool>, bool, usize, bool)single_line_if_else_max_width: (Cell<bool>, bool, usize, bool)single_line_let_else_max_width: (Cell<bool>, bool, usize, bool)wrap_comments: (Cell<bool>, bool, bool, bool)format_code_in_doc_comments: (Cell<bool>, bool, bool, bool)doc_comment_code_block_width: (Cell<bool>, bool, usize, bool)comment_width: (Cell<bool>, bool, usize, bool)normalize_comments: (Cell<bool>, bool, bool, bool)normalize_doc_attributes: (Cell<bool>, bool, bool, bool)format_strings: (Cell<bool>, bool, bool, bool)format_macro_matchers: (Cell<bool>, bool, bool, bool)format_macro_bodies: (Cell<bool>, bool, bool, bool)skip_macro_invocations: (Cell<bool>, bool, MacroSelectors, bool)hex_literal_case: (Cell<bool>, bool, HexLiteralCase, bool)empty_item_single_line: (Cell<bool>, bool, bool, bool)struct_lit_single_line: (Cell<bool>, bool, bool, bool)fn_single_line: (Cell<bool>, bool, bool, bool)where_single_line: (Cell<bool>, bool, bool, bool)imports_indent: (Cell<bool>, bool, IndentStyle, bool)imports_layout: (Cell<bool>, bool, ListTactic, bool)imports_granularity: (Cell<bool>, bool, ImportGranularity, bool)group_imports: (Cell<bool>, bool, GroupImportsTactic, bool)merge_imports: (Cell<bool>, bool, bool, bool)reorder_imports: (Cell<bool>, bool, bool, bool)reorder_modules: (Cell<bool>, bool, bool, bool)reorder_impl_items: (Cell<bool>, bool, bool, bool)type_punctuation_density: (Cell<bool>, bool, TypeDensity, bool)space_before_colon: (Cell<bool>, bool, bool, bool)space_after_colon: (Cell<bool>, bool, bool, bool)spaces_around_ranges: (Cell<bool>, bool, bool, bool)binop_separator: (Cell<bool>, bool, SeparatorPlace, bool)remove_nested_parens: (Cell<bool>, bool, bool, bool)combine_control_expr: (Cell<bool>, bool, bool, bool)short_array_element_width_threshold: (Cell<bool>, bool, usize, bool)overflow_delimited_expr: (Cell<bool>, bool, bool, bool)struct_field_align_threshold: (Cell<bool>, bool, usize, bool)enum_discrim_align_threshold: (Cell<bool>, bool, usize, bool)match_arm_blocks: (Cell<bool>, bool, bool, bool)match_arm_leading_pipes: (Cell<bool>, bool, MatchArmLeadingPipe, bool)force_multiline_blocks: (Cell<bool>, bool, bool, bool)fn_args_layout: (Cell<bool>, bool, Density, bool)fn_params_layout: (Cell<bool>, bool, Density, bool)brace_style: (Cell<bool>, bool, BraceStyle, bool)control_brace_style: (Cell<bool>, bool, ControlBraceStyle, bool)trailing_semicolon: (Cell<bool>, bool, bool, bool)trailing_comma: (Cell<bool>, bool, SeparatorTactic, bool)match_block_trailing_comma: (Cell<bool>, bool, bool, bool)blank_lines_upper_bound: (Cell<bool>, bool, usize, bool)blank_lines_lower_bound: (Cell<bool>, bool, usize, bool)edition: (Cell<bool>, bool, Edition, bool)version: (Cell<bool>, bool, Version, bool)inline_attribute_width: (Cell<bool>, bool, usize, bool)format_generated_files: (Cell<bool>, bool, bool, bool)merge_derives: (Cell<bool>, bool, bool, bool)use_try_shorthand: (Cell<bool>, bool, bool, bool)use_field_init_shorthand: (Cell<bool>, bool, bool, bool)force_explicit_abi: (Cell<bool>, bool, bool, bool)condense_wildcard_suffixes: (Cell<bool>, bool, bool, bool)color: (Cell<bool>, bool, Color, bool)required_version: (Cell<bool>, bool, String, bool)unstable_features: (Cell<bool>, bool, bool, bool)disable_all_formatting: (Cell<bool>, bool, bool, bool)skip_children: (Cell<bool>, bool, bool, bool)hide_parse_errors: (Cell<bool>, bool, bool, bool)error_on_line_overflow: (Cell<bool>, bool, bool, bool)error_on_unformatted: (Cell<bool>, bool, bool, bool)ignore: (Cell<bool>, bool, IgnoreList, bool)verbose: (Cell<bool>, bool, Verbosity, bool)file_lines: (Cell<bool>, bool, FileLines, bool)emit_mode: (Cell<bool>, bool, EmitMode, bool)make_backup: (Cell<bool>, bool, bool, bool)print_misformatted_file_names: (Cell<bool>, bool, bool, bool)
Implementations

impl Config

pub fn max_width(&self) -> usize

pub fn hard_tabs(&self) -> bool

pub fn tab_spaces(&self) -> usize

pub fn newline_style(&self) -> NewlineStyle

pub fn indent_style(&self) -> IndentStyle

pub fn use_small_heuristics(&self) -> Heuristics

pub fn width_heuristics(&self) -> WidthHeuristics

pub fn fn_call_width(&self) -> usize

pub fn attr_fn_like_width(&self) -> usize

pub fn struct_lit_width(&self) -> usize

pub fn struct_variant_width(&self) -> usize

pub fn array_width(&self) -> usize

pub fn chain_width(&self) -> usize

pub fn single_line_if_else_max_width(&self) -> usize

pub fn single_line_let_else_max_width(&self) -> usize

pub fn wrap_comments(&self) -> bool

pub fn format_code_in_doc_comments(&self) -> bool

pub fn doc_comment_code_block_width(&self) -> usize

pub fn comment_width(&self) -> usize

pub fn normalize_comments(&self) -> bool

pub fn normalize_doc_attributes(&self) -> bool

pub fn format_strings(&self) -> bool

pub fn format_macro_matchers(&self) -> bool

pub fn format_macro_bodies(&self) -> bool

pub fn skip_macro_invocations(&self) -> MacroSelectors

pub fn hex_literal_case(&self) -> HexLiteralCase

pub fn empty_item_single_line(&self) -> bool

pub fn struct_lit_single_line(&self) -> bool

pub fn fn_single_line(&self) -> bool

pub fn where_single_line(&self) -> bool

pub fn imports_indent(&self) -> IndentStyle

pub fn imports_layout(&self) -> ListTactic

pub fn imports_granularity(&self) -> ImportGranularity

pub fn group_imports(&self) -> GroupImportsTactic

pub fn merge_imports(&self) -> bool

pub fn reorder_imports(&self) -> bool

pub fn reorder_modules(&self) -> bool

pub fn reorder_impl_items(&self) -> bool

pub fn type_punctuation_density(&self) -> TypeDensity

pub fn space_before_colon(&self) -> bool

pub fn space_after_colon(&self) -> bool

pub fn spaces_around_ranges(&self) -> bool

pub fn binop_separator(&self) -> SeparatorPlace

pub fn remove_nested_parens(&self) -> bool

pub fn combine_control_expr(&self) -> bool

pub fn short_array_element_width_threshold(&self) -> usize

pub fn overflow_delimited_expr(&self) -> bool

pub fn struct_field_align_threshold(&self) -> usize

pub fn enum_discrim_align_threshold(&self) -> usize

pub fn match_arm_blocks(&self) -> bool

pub fn match_arm_leading_pipes(&self) -> MatchArmLeadingPipe

pub fn force_multiline_blocks(&self) -> bool

pub fn fn_args_layout(&self) -> Density

pub fn fn_params_layout(&self) -> Density

pub fn brace_style(&self) -> BraceStyle

pub fn control_brace_style(&self) -> ControlBraceStyle

pub fn trailing_semicolon(&self) -> bool

pub fn trailing_comma(&self) -> SeparatorTactic

pub fn match_block_trailing_comma(&self) -> bool

pub fn blank_lines_upper_bound(&self) -> usize

pub fn blank_lines_lower_bound(&self) -> usize

pub fn edition(&self) -> Edition

pub fn version(&self) -> Version

pub fn inline_attribute_width(&self) -> usize

pub fn format_generated_files(&self) -> bool

pub fn merge_derives(&self) -> bool

pub fn use_try_shorthand(&self) -> bool

pub fn use_field_init_shorthand(&self) -> bool

pub fn force_explicit_abi(&self) -> bool

pub fn condense_wildcard_suffixes(&self) -> bool

pub fn color(&self) -> Color

pub fn required_version(&self) -> String

pub fn unstable_features(&self) -> bool

pub fn disable_all_formatting(&self) -> bool

pub fn skip_children(&self) -> bool

pub fn hide_parse_errors(&self) -> bool

pub fn error_on_line_overflow(&self) -> bool

pub fn error_on_unformatted(&self) -> bool

pub fn ignore(&self) -> IgnoreList

pub fn verbose(&self) -> Verbosity

pub fn file_lines(&self) -> FileLines

pub fn emit_mode(&self) -> EmitMode

pub fn make_backup(&self) -> bool

pub fn print_misformatted_file_names(&self) -> bool

pub fn set(&mut self) -> ConfigSetter<'_>

pub fn was_set(&self) -> ConfigWasSet<'_>

fn fill_from_parsed_config(self, parsed: PartialConfig, dir: &Path) -> Config

pub(crate) fn is_valid_name(name: &str) -> bool

pub fn is_valid_key_val(key: &str, val: &str) -> bool

pub fn used_options(&self) -> PartialConfig

pub fn all_options(&self) -> PartialConfig

pub fn override_value(&mut self, key: &str, val: &str)

pub fn is_hidden_option(name: &str) -> bool

pub fn print_docs(out: &mut dyn Write, include_unstable: bool)

fn set_width_heuristics(&mut self, heuristics: WidthHeuristics)

fn set_heuristics(&mut self)

fn set_ignore(&mut self, dir: &Path)

fn set_merge_imports(&mut self)

fn set_fn_args_layout(&mut self)

pub fn is_default(&self, key: &str) -> bool

Returns true if the config key was explicitly set and is the default value.

impl Config

pub(crate) fn version_meets_requirement(&self) -> bool

pub(crate) fn from_toml_path(file_path: &Path) -> Result<Config, Error>

Constructs a Config from the toml file specified at file_path.

This method only looks at the provided path, for a method that searches parents for a rustfmt.toml see from_resolved_toml_path.

Returns a Config if the config could be read and parsed from the file, otherwise errors.

pub(crate) fn from_resolved_toml_path( dir: &Path ) -> Result<(Config, Option<PathBuf>), Error>

Resolves the config for input in dir.

Searches for rustfmt.toml beginning with dir, and recursively checking parents of dir if no config file is found. If no config file exists in dir or in any parent, a default Config will be returned (and the returned path will be empty).

Returns the Config to use, and the path of the project file if there was one.

pub(crate) fn from_toml(toml: &str, dir: &Path) -> Result<Config, String>