rustfmt_nightly::config
Struct Config

pub struct Config {

    max_width: (Cell<bool>, bool, <MaxWidth as StyleEditionDefault>::ConfigType, bool, bool),
    hard_tabs: (Cell<bool>, bool, <HardTabs as StyleEditionDefault>::ConfigType, bool, bool),
    tab_spaces: (Cell<bool>, bool, <TabSpaces as StyleEditionDefault>::ConfigType, bool, bool),
    newline_style: (Cell<bool>, bool, <NewlineStyleConfig as StyleEditionDefault>::ConfigType, bool, bool),
    indent_style: (Cell<bool>, bool, <IndentStyleConfig as StyleEditionDefault>::ConfigType, bool, bool),
    use_small_heuristics: (Cell<bool>, bool, <UseSmallHeuristics as StyleEditionDefault>::ConfigType, bool, bool),
    width_heuristics: (Cell<bool>, bool, <WidthHeuristicsConfig as StyleEditionDefault>::ConfigType, bool, bool),
    fn_call_width: (Cell<bool>, bool, <FnCallWidth as StyleEditionDefault>::ConfigType, bool, bool),
    attr_fn_like_width: (Cell<bool>, bool, <AttrFnLikeWidth as StyleEditionDefault>::ConfigType, bool, bool),
    struct_lit_width: (Cell<bool>, bool, <StructLitWidth as StyleEditionDefault>::ConfigType, bool, bool),
    struct_variant_width: (Cell<bool>, bool, <StructVariantWidth as StyleEditionDefault>::ConfigType, bool, bool),
    array_width: (Cell<bool>, bool, <ArrayWidth as StyleEditionDefault>::ConfigType, bool, bool),
    chain_width: (Cell<bool>, bool, <ChainWidth as StyleEditionDefault>::ConfigType, bool, bool),
    single_line_if_else_max_width: (Cell<bool>, bool, <SingleLineIfElseMaxWidth as StyleEditionDefault>::ConfigType, bool, bool),
    single_line_let_else_max_width: (Cell<bool>, bool, <SingleLineLetElseMaxWidth as StyleEditionDefault>::ConfigType, bool, bool),
    wrap_comments: (Cell<bool>, bool, <WrapComments as StyleEditionDefault>::ConfigType, bool, bool),
    format_code_in_doc_comments: (Cell<bool>, bool, <FormatCodeInDocComments as StyleEditionDefault>::ConfigType, bool, bool),
    doc_comment_code_block_width: (Cell<bool>, bool, <DocCommentCodeBlockWidth as StyleEditionDefault>::ConfigType, bool, bool),
    comment_width: (Cell<bool>, bool, <CommentWidth as StyleEditionDefault>::ConfigType, bool, bool),
    normalize_comments: (Cell<bool>, bool, <NormalizeComments as StyleEditionDefault>::ConfigType, bool, bool),
    normalize_doc_attributes: (Cell<bool>, bool, <NormalizeDocAttributes as StyleEditionDefault>::ConfigType, bool, bool),
    format_strings: (Cell<bool>, bool, <FormatStrings as StyleEditionDefault>::ConfigType, bool, bool),
    format_macro_matchers: (Cell<bool>, bool, <FormatMacroMatchers as StyleEditionDefault>::ConfigType, bool, bool),
    format_macro_bodies: (Cell<bool>, bool, <FormatMacroBodies as StyleEditionDefault>::ConfigType, bool, bool),
    skip_macro_invocations: (Cell<bool>, bool, <SkipMacroInvocations as StyleEditionDefault>::ConfigType, bool, bool),
    hex_literal_case: (Cell<bool>, bool, <HexLiteralCaseConfig as StyleEditionDefault>::ConfigType, bool, bool),
    empty_item_single_line: (Cell<bool>, bool, <EmptyItemSingleLine as StyleEditionDefault>::ConfigType, bool, bool),
    struct_lit_single_line: (Cell<bool>, bool, <StructLitSingleLine as StyleEditionDefault>::ConfigType, bool, bool),
    fn_single_line: (Cell<bool>, bool, <FnSingleLine as StyleEditionDefault>::ConfigType, bool, bool),
    where_single_line: (Cell<bool>, bool, <WhereSingleLine as StyleEditionDefault>::ConfigType, bool, bool),
    imports_indent: (Cell<bool>, bool, <ImportsIndent as StyleEditionDefault>::ConfigType, bool, bool),
    imports_layout: (Cell<bool>, bool, <ImportsLayout as StyleEditionDefault>::ConfigType, bool, bool),
    imports_granularity: (Cell<bool>, bool, <ImportsGranularityConfig as StyleEditionDefault>::ConfigType, bool, bool),
    group_imports: (Cell<bool>, bool, <GroupImportsTacticConfig as StyleEditionDefault>::ConfigType, bool, bool),
    merge_imports: (Cell<bool>, bool, <MergeImports as StyleEditionDefault>::ConfigType, bool, bool),
    reorder_imports: (Cell<bool>, bool, <ReorderImports as StyleEditionDefault>::ConfigType, bool, bool),
    reorder_modules: (Cell<bool>, bool, <ReorderModules as StyleEditionDefault>::ConfigType, bool, bool),
    reorder_impl_items: (Cell<bool>, bool, <ReorderImplItems as StyleEditionDefault>::ConfigType, bool, bool),
    type_punctuation_density: (Cell<bool>, bool, <TypePunctuationDensity as StyleEditionDefault>::ConfigType, bool, bool),
    space_before_colon: (Cell<bool>, bool, <SpaceBeforeColon as StyleEditionDefault>::ConfigType, bool, bool),
    space_after_colon: (Cell<bool>, bool, <SpaceAfterColon as StyleEditionDefault>::ConfigType, bool, bool),
    spaces_around_ranges: (Cell<bool>, bool, <SpacesAroundRanges as StyleEditionDefault>::ConfigType, bool, bool),
    binop_separator: (Cell<bool>, bool, <BinopSeparator as StyleEditionDefault>::ConfigType, bool, bool),
    remove_nested_parens: (Cell<bool>, bool, <RemoveNestedParens as StyleEditionDefault>::ConfigType, bool, bool),
    combine_control_expr: (Cell<bool>, bool, <CombineControlExpr as StyleEditionDefault>::ConfigType, bool, bool),
    short_array_element_width_threshold: (Cell<bool>, bool, <ShortArrayElementWidthThreshold as StyleEditionDefault>::ConfigType, bool, bool),
    overflow_delimited_expr: (Cell<bool>, bool, <OverflowDelimitedExpr as StyleEditionDefault>::ConfigType, bool, bool),
    struct_field_align_threshold: (Cell<bool>, bool, <StructFieldAlignThreshold as StyleEditionDefault>::ConfigType, bool, bool),
    enum_discrim_align_threshold: (Cell<bool>, bool, <EnumDiscrimAlignThreshold as StyleEditionDefault>::ConfigType, bool, bool),
    match_arm_blocks: (Cell<bool>, bool, <MatchArmBlocks as StyleEditionDefault>::ConfigType, bool, bool),
    match_arm_leading_pipes: (Cell<bool>, bool, <MatchArmLeadingPipeConfig as StyleEditionDefault>::ConfigType, bool, bool),
    force_multiline_blocks: (Cell<bool>, bool, <ForceMultilineBlocks as StyleEditionDefault>::ConfigType, bool, bool),
    fn_args_layout: (Cell<bool>, bool, <FnArgsLayout as StyleEditionDefault>::ConfigType, bool, bool),
    fn_params_layout: (Cell<bool>, bool, <FnParamsLayout as StyleEditionDefault>::ConfigType, bool, bool),
    brace_style: (Cell<bool>, bool, <BraceStyleConfig as StyleEditionDefault>::ConfigType, bool, bool),
    control_brace_style: (Cell<bool>, bool, <ControlBraceStyleConfig as StyleEditionDefault>::ConfigType, bool, bool),
    trailing_semicolon: (Cell<bool>, bool, <TrailingSemicolon as StyleEditionDefault>::ConfigType, bool, bool),
    trailing_comma: (Cell<bool>, bool, <TrailingComma as StyleEditionDefault>::ConfigType, bool, bool),
    match_block_trailing_comma: (Cell<bool>, bool, <MatchBlockTrailingComma as StyleEditionDefault>::ConfigType, bool, bool),
    blank_lines_upper_bound: (Cell<bool>, bool, <BlankLinesUpperBound as StyleEditionDefault>::ConfigType, bool, bool),
    blank_lines_lower_bound: (Cell<bool>, bool, <BlankLinesLowerBound as StyleEditionDefault>::ConfigType, bool, bool),
    edition: (Cell<bool>, bool, <EditionConfig as StyleEditionDefault>::ConfigType, bool, bool),
    style_edition: (Cell<bool>, bool, <StyleEditionConfig as StyleEditionDefault>::ConfigType, bool, bool),
    version: (Cell<bool>, bool, <VersionConfig as StyleEditionDefault>::ConfigType, bool, bool),
    inline_attribute_width: (Cell<bool>, bool, <InlineAttributeWidth as StyleEditionDefault>::ConfigType, bool, bool),
    format_generated_files: (Cell<bool>, bool, <FormatGeneratedFiles as StyleEditionDefault>::ConfigType, bool, bool),
    generated_marker_line_search_limit: (Cell<bool>, bool, <GeneratedMarkerLineSearchLimit as StyleEditionDefault>::ConfigType, bool, bool),
    merge_derives: (Cell<bool>, bool, <MergeDerives as StyleEditionDefault>::ConfigType, bool, bool),
    use_try_shorthand: (Cell<bool>, bool, <UseTryShorthand as StyleEditionDefault>::ConfigType, bool, bool),
    use_field_init_shorthand: (Cell<bool>, bool, <UseFieldInitShorthand as StyleEditionDefault>::ConfigType, bool, bool),
    force_explicit_abi: (Cell<bool>, bool, <ForceExplicitAbi as StyleEditionDefault>::ConfigType, bool, bool),
    condense_wildcard_suffixes: (Cell<bool>, bool, <CondenseWildcardSuffixes as StyleEditionDefault>::ConfigType, bool, bool),
    color: (Cell<bool>, bool, <ColorConfig as StyleEditionDefault>::ConfigType, bool, bool),
    required_version: (Cell<bool>, bool, <RequiredVersion as StyleEditionDefault>::ConfigType, bool, bool),
    unstable_features: (Cell<bool>, bool, <UnstableFeatures as StyleEditionDefault>::ConfigType, bool, bool),
    disable_all_formatting: (Cell<bool>, bool, <DisableAllFormatting as StyleEditionDefault>::ConfigType, bool, bool),
    skip_children: (Cell<bool>, bool, <SkipChildren as StyleEditionDefault>::ConfigType, bool, bool),
    hide_parse_errors: (Cell<bool>, bool, <HideParseErrors as StyleEditionDefault>::ConfigType, bool, bool),
    show_parse_errors: (Cell<bool>, bool, <ShowParseErrors as StyleEditionDefault>::ConfigType, bool, bool),
    error_on_line_overflow: (Cell<bool>, bool, <ErrorOnLineOverflow as StyleEditionDefault>::ConfigType, bool, bool),
    error_on_unformatted: (Cell<bool>, bool, <ErrorOnUnformatted as StyleEditionDefault>::ConfigType, bool, bool),
    ignore: (Cell<bool>, bool, <Ignore as StyleEditionDefault>::ConfigType, bool, bool),
    verbose: (Cell<bool>, bool, <Verbose as StyleEditionDefault>::ConfigType, bool, bool),
    file_lines: (Cell<bool>, bool, <FileLinesConfig as StyleEditionDefault>::ConfigType, bool, bool),
    emit_mode: (Cell<bool>, bool, <EmitModeConfig as StyleEditionDefault>::ConfigType, bool, bool),
    make_backup: (Cell<bool>, bool, <MakeBackup as StyleEditionDefault>::ConfigType, bool, bool),
    print_misformatted_file_names: (Cell<bool>, bool, <PrintMisformattedFileNames as StyleEditionDefault>::ConfigType, bool, bool),

}

Fields
max_width: (Cell<bool>, bool, <MaxWidth as StyleEditionDefault>::ConfigType, bool, bool)hard_tabs: (Cell<bool>, bool, <HardTabs as StyleEditionDefault>::ConfigType, bool, bool)tab_spaces: (Cell<bool>, bool, <TabSpaces as StyleEditionDefault>::ConfigType, bool, bool)newline_style: (Cell<bool>, bool, <NewlineStyleConfig as StyleEditionDefault>::ConfigType, bool, bool)indent_style: (Cell<bool>, bool, <IndentStyleConfig as StyleEditionDefault>::ConfigType, bool, bool)use_small_heuristics: (Cell<bool>, bool, <UseSmallHeuristics as StyleEditionDefault>::ConfigType, bool, bool)width_heuristics: (Cell<bool>, bool, <WidthHeuristicsConfig as StyleEditionDefault>::ConfigType, bool, bool)fn_call_width: (Cell<bool>, bool, <FnCallWidth as StyleEditionDefault>::ConfigType, bool, bool)attr_fn_like_width: (Cell<bool>, bool, <AttrFnLikeWidth as StyleEditionDefault>::ConfigType, bool, bool)struct_lit_width: (Cell<bool>, bool, <StructLitWidth as StyleEditionDefault>::ConfigType, bool, bool)struct_variant_width: (Cell<bool>, bool, <StructVariantWidth as StyleEditionDefault>::ConfigType, bool, bool)array_width: (Cell<bool>, bool, <ArrayWidth as StyleEditionDefault>::ConfigType, bool, bool)chain_width: (Cell<bool>, bool, <ChainWidth as StyleEditionDefault>::ConfigType, bool, bool)single_line_if_else_max_width: (Cell<bool>, bool, <SingleLineIfElseMaxWidth as StyleEditionDefault>::ConfigType, bool, bool)single_line_let_else_max_width: (Cell<bool>, bool, <SingleLineLetElseMaxWidth as StyleEditionDefault>::ConfigType, bool, bool)wrap_comments: (Cell<bool>, bool, <WrapComments as StyleEditionDefault>::ConfigType, bool, bool)format_code_in_doc_comments: (Cell<bool>, bool, <FormatCodeInDocComments as StyleEditionDefault>::ConfigType, bool, bool)doc_comment_code_block_width: (Cell<bool>, bool, <DocCommentCodeBlockWidth as StyleEditionDefault>::ConfigType, bool, bool)comment_width: (Cell<bool>, bool, <CommentWidth as StyleEditionDefault>::ConfigType, bool, bool)normalize_comments: (Cell<bool>, bool, <NormalizeComments as StyleEditionDefault>::ConfigType, bool, bool)normalize_doc_attributes: (Cell<bool>, bool, <NormalizeDocAttributes as StyleEditionDefault>::ConfigType, bool, bool)format_strings: (Cell<bool>, bool, <FormatStrings as StyleEditionDefault>::ConfigType, bool, bool)format_macro_matchers: (Cell<bool>, bool, <FormatMacroMatchers as StyleEditionDefault>::ConfigType, bool, bool)format_macro_bodies: (Cell<bool>, bool, <FormatMacroBodies as StyleEditionDefault>::ConfigType, bool, bool)skip_macro_invocations: (Cell<bool>, bool, <SkipMacroInvocations as StyleEditionDefault>::ConfigType, bool, bool)hex_literal_case: (Cell<bool>, bool, <HexLiteralCaseConfig as StyleEditionDefault>::ConfigType, bool, bool)empty_item_single_line: (Cell<bool>, bool, <EmptyItemSingleLine as StyleEditionDefault>::ConfigType, bool, bool)struct_lit_single_line: (Cell<bool>, bool, <StructLitSingleLine as StyleEditionDefault>::ConfigType, bool, bool)fn_single_line: (Cell<bool>, bool, <FnSingleLine as StyleEditionDefault>::ConfigType, bool, bool)where_single_line: (Cell<bool>, bool, <WhereSingleLine as StyleEditionDefault>::ConfigType, bool, bool)imports_indent: (Cell<bool>, bool, <ImportsIndent as StyleEditionDefault>::ConfigType, bool, bool)imports_layout: (Cell<bool>, bool, <ImportsLayout as StyleEditionDefault>::ConfigType, bool, bool)imports_granularity: (Cell<bool>, bool, <ImportsGranularityConfig as StyleEditionDefault>::ConfigType, bool, bool)group_imports: (Cell<bool>, bool, <GroupImportsTacticConfig as StyleEditionDefault>::ConfigType, bool, bool)merge_imports: (Cell<bool>, bool, <MergeImports as StyleEditionDefault>::ConfigType, bool, bool)reorder_imports: (Cell<bool>, bool, <ReorderImports as StyleEditionDefault>::ConfigType, bool, bool)reorder_modules: (Cell<bool>, bool, <ReorderModules as StyleEditionDefault>::ConfigType, bool, bool)reorder_impl_items: (Cell<bool>, bool, <ReorderImplItems as StyleEditionDefault>::ConfigType, bool, bool)type_punctuation_density: (Cell<bool>, bool, <TypePunctuationDensity as StyleEditionDefault>::ConfigType, bool, bool)space_before_colon: (Cell<bool>, bool, <SpaceBeforeColon as StyleEditionDefault>::ConfigType, bool, bool)space_after_colon: (Cell<bool>, bool, <SpaceAfterColon as StyleEditionDefault>::ConfigType, bool, bool)spaces_around_ranges: (Cell<bool>, bool, <SpacesAroundRanges as StyleEditionDefault>::ConfigType, bool, bool)binop_separator: (Cell<bool>, bool, <BinopSeparator as StyleEditionDefault>::ConfigType, bool, bool)remove_nested_parens: (Cell<bool>, bool, <RemoveNestedParens as StyleEditionDefault>::ConfigType, bool, bool)combine_control_expr: (Cell<bool>, bool, <CombineControlExpr as StyleEditionDefault>::ConfigType, bool, bool)short_array_element_width_threshold: (Cell<bool>, bool, <ShortArrayElementWidthThreshold as StyleEditionDefault>::ConfigType, bool, bool)overflow_delimited_expr: (Cell<bool>, bool, <OverflowDelimitedExpr as StyleEditionDefault>::ConfigType, bool, bool)struct_field_align_threshold: (Cell<bool>, bool, <StructFieldAlignThreshold as StyleEditionDefault>::ConfigType, bool, bool)enum_discrim_align_threshold: (Cell<bool>, bool, <EnumDiscrimAlignThreshold as StyleEditionDefault>::ConfigType, bool, bool)match_arm_blocks: (Cell<bool>, bool, <MatchArmBlocks as StyleEditionDefault>::ConfigType, bool, bool)match_arm_leading_pipes: (Cell<bool>, bool, <MatchArmLeadingPipeConfig as StyleEditionDefault>::ConfigType, bool, bool)force_multiline_blocks: (Cell<bool>, bool, <ForceMultilineBlocks as StyleEditionDefault>::ConfigType, bool, bool)fn_args_layout: (Cell<bool>, bool, <FnArgsLayout as StyleEditionDefault>::ConfigType, bool, bool)fn_params_layout: (Cell<bool>, bool, <FnParamsLayout as StyleEditionDefault>::ConfigType, bool, bool)brace_style: (Cell<bool>, bool, <BraceStyleConfig as StyleEditionDefault>::ConfigType, bool, bool)control_brace_style: (Cell<bool>, bool, <ControlBraceStyleConfig as StyleEditionDefault>::ConfigType, bool, bool)trailing_semicolon: (Cell<bool>, bool, <TrailingSemicolon as StyleEditionDefault>::ConfigType, bool, bool)trailing_comma: (Cell<bool>, bool, <TrailingComma as StyleEditionDefault>::ConfigType, bool, bool)match_block_trailing_comma: (Cell<bool>, bool, <MatchBlockTrailingComma as StyleEditionDefault>::ConfigType, bool, bool)blank_lines_upper_bound: (Cell<bool>, bool, <BlankLinesUpperBound as StyleEditionDefault>::ConfigType, bool, bool)blank_lines_lower_bound: (Cell<bool>, bool, <BlankLinesLowerBound as StyleEditionDefault>::ConfigType, bool, bool)edition: (Cell<bool>, bool, <EditionConfig as StyleEditionDefault>::ConfigType, bool, bool)style_edition: (Cell<bool>, bool, <StyleEditionConfig as StyleEditionDefault>::ConfigType, bool, bool)version: (Cell<bool>, bool, <VersionConfig as StyleEditionDefault>::ConfigType, bool, bool)inline_attribute_width: (Cell<bool>, bool, <InlineAttributeWidth as StyleEditionDefault>::ConfigType, bool, bool)format_generated_files: (Cell<bool>, bool, <FormatGeneratedFiles as StyleEditionDefault>::ConfigType, bool, bool)generated_marker_line_search_limit: (Cell<bool>, bool, <GeneratedMarkerLineSearchLimit as StyleEditionDefault>::ConfigType, bool, bool)merge_derives: (Cell<bool>, bool, <MergeDerives as StyleEditionDefault>::ConfigType, bool, bool)use_try_shorthand: (Cell<bool>, bool, <UseTryShorthand as StyleEditionDefault>::ConfigType, bool, bool)use_field_init_shorthand: (Cell<bool>, bool, <UseFieldInitShorthand as StyleEditionDefault>::ConfigType, bool, bool)force_explicit_abi: (Cell<bool>, bool, <ForceExplicitAbi as StyleEditionDefault>::ConfigType, bool, bool)condense_wildcard_suffixes: (Cell<bool>, bool, <CondenseWildcardSuffixes as StyleEditionDefault>::ConfigType, bool, bool)color: (Cell<bool>, bool, <ColorConfig as StyleEditionDefault>::ConfigType, bool, bool)required_version: (Cell<bool>, bool, <RequiredVersion as StyleEditionDefault>::ConfigType, bool, bool)unstable_features: (Cell<bool>, bool, <UnstableFeatures as StyleEditionDefault>::ConfigType, bool, bool)disable_all_formatting: (Cell<bool>, bool, <DisableAllFormatting as StyleEditionDefault>::ConfigType, bool, bool)skip_children: (Cell<bool>, bool, <SkipChildren as StyleEditionDefault>::ConfigType, bool, bool)hide_parse_errors: (Cell<bool>, bool, <HideParseErrors as StyleEditionDefault>::ConfigType, bool, bool)show_parse_errors: (Cell<bool>, bool, <ShowParseErrors as StyleEditionDefault>::ConfigType, bool, bool)error_on_line_overflow: (Cell<bool>, bool, <ErrorOnLineOverflow as StyleEditionDefault>::ConfigType, bool, bool)error_on_unformatted: (Cell<bool>, bool, <ErrorOnUnformatted as StyleEditionDefault>::ConfigType, bool, bool)ignore: (Cell<bool>, bool, <Ignore as StyleEditionDefault>::ConfigType, bool, bool)verbose: (Cell<bool>, bool, <Verbose as StyleEditionDefault>::ConfigType, bool, bool)file_lines: (Cell<bool>, bool, <FileLinesConfig as StyleEditionDefault>::ConfigType, bool, bool)emit_mode: (Cell<bool>, bool, <EmitModeConfig as StyleEditionDefault>::ConfigType, bool, bool)make_backup: (Cell<bool>, bool, <MakeBackup as StyleEditionDefault>::ConfigType, bool, bool)print_misformatted_file_names: (Cell<bool>, bool, <PrintMisformattedFileNames as StyleEditionDefault>::ConfigType, bool, bool)
Implementations

impl Config

pub fn max_width(&self) -> <MaxWidth as StyleEditionDefault>::ConfigType

pub fn hard_tabs(&self) -> <HardTabs as StyleEditionDefault>::ConfigType

pub fn tab_spaces(&self) -> <TabSpaces as StyleEditionDefault>::ConfigType

pub fn newline_style( &self, ) -> <NewlineStyleConfig as StyleEditionDefault>::ConfigType

pub fn indent_style( &self, ) -> <IndentStyleConfig as StyleEditionDefault>::ConfigType

pub fn use_small_heuristics( &self, ) -> <UseSmallHeuristics as StyleEditionDefault>::ConfigType

pub fn width_heuristics( &self, ) -> <WidthHeuristicsConfig as StyleEditionDefault>::ConfigType

pub fn fn_call_width(&self) -> <FnCallWidth as StyleEditionDefault>::ConfigType

pub fn attr_fn_like_width( &self, ) -> <AttrFnLikeWidth as StyleEditionDefault>::ConfigType

pub fn struct_lit_width( &self, ) -> <StructLitWidth as StyleEditionDefault>::ConfigType

pub fn struct_variant_width( &self, ) -> <StructVariantWidth as StyleEditionDefault>::ConfigType

pub fn array_width(&self) -> <ArrayWidth as StyleEditionDefault>::ConfigType

pub fn chain_width(&self) -> <ChainWidth as StyleEditionDefault>::ConfigType

pub fn single_line_if_else_max_width( &self, ) -> <SingleLineIfElseMaxWidth as StyleEditionDefault>::ConfigType

pub fn single_line_let_else_max_width( &self, ) -> <SingleLineLetElseMaxWidth as StyleEditionDefault>::ConfigType

pub fn wrap_comments(&self) -> <WrapComments as StyleEditionDefault>::ConfigType

pub fn format_code_in_doc_comments( &self, ) -> <FormatCodeInDocComments as StyleEditionDefault>::ConfigType

pub fn doc_comment_code_block_width( &self, ) -> <DocCommentCodeBlockWidth as StyleEditionDefault>::ConfigType

pub fn comment_width(&self) -> <CommentWidth as StyleEditionDefault>::ConfigType

pub fn normalize_comments( &self, ) -> <NormalizeComments as StyleEditionDefault>::ConfigType

pub fn normalize_doc_attributes( &self, ) -> <NormalizeDocAttributes as StyleEditionDefault>::ConfigType

pub fn format_strings( &self, ) -> <FormatStrings as StyleEditionDefault>::ConfigType

pub fn format_macro_matchers( &self, ) -> <FormatMacroMatchers as StyleEditionDefault>::ConfigType

pub fn format_macro_bodies( &self, ) -> <FormatMacroBodies as StyleEditionDefault>::ConfigType

pub fn skip_macro_invocations( &self, ) -> <SkipMacroInvocations as StyleEditionDefault>::ConfigType

pub fn hex_literal_case( &self, ) -> <HexLiteralCaseConfig as StyleEditionDefault>::ConfigType

pub fn empty_item_single_line( &self, ) -> <EmptyItemSingleLine as StyleEditionDefault>::ConfigType

pub fn struct_lit_single_line( &self, ) -> <StructLitSingleLine as StyleEditionDefault>::ConfigType

pub fn fn_single_line( &self, ) -> <FnSingleLine as StyleEditionDefault>::ConfigType

pub fn where_single_line( &self, ) -> <WhereSingleLine as StyleEditionDefault>::ConfigType

pub fn imports_indent( &self, ) -> <ImportsIndent as StyleEditionDefault>::ConfigType

pub fn imports_layout( &self, ) -> <ImportsLayout as StyleEditionDefault>::ConfigType

pub fn imports_granularity( &self, ) -> <ImportsGranularityConfig as StyleEditionDefault>::ConfigType

pub fn group_imports( &self, ) -> <GroupImportsTacticConfig as StyleEditionDefault>::ConfigType

pub fn merge_imports(&self) -> <MergeImports as StyleEditionDefault>::ConfigType

pub fn reorder_imports( &self, ) -> <ReorderImports as StyleEditionDefault>::ConfigType

pub fn reorder_modules( &self, ) -> <ReorderModules as StyleEditionDefault>::ConfigType

pub fn reorder_impl_items( &self, ) -> <ReorderImplItems as StyleEditionDefault>::ConfigType

pub fn type_punctuation_density( &self, ) -> <TypePunctuationDensity as StyleEditionDefault>::ConfigType

pub fn space_before_colon( &self, ) -> <SpaceBeforeColon as StyleEditionDefault>::ConfigType

pub fn space_after_colon( &self, ) -> <SpaceAfterColon as StyleEditionDefault>::ConfigType

pub fn spaces_around_ranges( &self, ) -> <SpacesAroundRanges as StyleEditionDefault>::ConfigType

pub fn binop_separator( &self, ) -> <BinopSeparator as StyleEditionDefault>::ConfigType

pub fn remove_nested_parens( &self, ) -> <RemoveNestedParens as StyleEditionDefault>::ConfigType

pub fn combine_control_expr( &self, ) -> <CombineControlExpr as StyleEditionDefault>::ConfigType

pub fn short_array_element_width_threshold( &self, ) -> <ShortArrayElementWidthThreshold as StyleEditionDefault>::ConfigType

pub fn overflow_delimited_expr( &self, ) -> <OverflowDelimitedExpr as StyleEditionDefault>::ConfigType

pub fn struct_field_align_threshold( &self, ) -> <StructFieldAlignThreshold as StyleEditionDefault>::ConfigType

pub fn enum_discrim_align_threshold( &self, ) -> <EnumDiscrimAlignThreshold as StyleEditionDefault>::ConfigType

pub fn match_arm_blocks( &self, ) -> <MatchArmBlocks as StyleEditionDefault>::ConfigType

pub fn match_arm_leading_pipes( &self, ) -> <MatchArmLeadingPipeConfig as StyleEditionDefault>::ConfigType

pub fn force_multiline_blocks( &self, ) -> <ForceMultilineBlocks as StyleEditionDefault>::ConfigType

pub fn fn_args_layout( &self, ) -> <FnArgsLayout as StyleEditionDefault>::ConfigType

pub fn fn_params_layout( &self, ) -> <FnParamsLayout as StyleEditionDefault>::ConfigType

pub fn brace_style( &self, ) -> <BraceStyleConfig as StyleEditionDefault>::ConfigType

pub fn control_brace_style( &self, ) -> <ControlBraceStyleConfig as StyleEditionDefault>::ConfigType

pub fn trailing_semicolon( &self, ) -> <TrailingSemicolon as StyleEditionDefault>::ConfigType

pub fn trailing_comma( &self, ) -> <TrailingComma as StyleEditionDefault>::ConfigType

pub fn match_block_trailing_comma( &self, ) -> <MatchBlockTrailingComma as StyleEditionDefault>::ConfigType

pub fn blank_lines_upper_bound( &self, ) -> <BlankLinesUpperBound as StyleEditionDefault>::ConfigType

pub fn blank_lines_lower_bound( &self, ) -> <BlankLinesLowerBound as StyleEditionDefault>::ConfigType

pub fn edition(&self) -> <EditionConfig as StyleEditionDefault>::ConfigType

pub fn style_edition( &self, ) -> <StyleEditionConfig as StyleEditionDefault>::ConfigType

pub fn version(&self) -> <VersionConfig as StyleEditionDefault>::ConfigType

pub fn inline_attribute_width( &self, ) -> <InlineAttributeWidth as StyleEditionDefault>::ConfigType

pub fn format_generated_files( &self, ) -> <FormatGeneratedFiles as StyleEditionDefault>::ConfigType

pub fn generated_marker_line_search_limit( &self, ) -> <GeneratedMarkerLineSearchLimit as StyleEditionDefault>::ConfigType

pub fn merge_derives(&self) -> <MergeDerives as StyleEditionDefault>::ConfigType

pub fn use_try_shorthand( &self, ) -> <UseTryShorthand as StyleEditionDefault>::ConfigType

pub fn use_field_init_shorthand( &self, ) -> <UseFieldInitShorthand as StyleEditionDefault>::ConfigType

pub fn force_explicit_abi( &self, ) -> <ForceExplicitAbi as StyleEditionDefault>::ConfigType

pub fn condense_wildcard_suffixes( &self, ) -> <CondenseWildcardSuffixes as StyleEditionDefault>::ConfigType

pub fn color(&self) -> <ColorConfig as StyleEditionDefault>::ConfigType

pub fn required_version( &self, ) -> <RequiredVersion as StyleEditionDefault>::ConfigType

pub fn unstable_features( &self, ) -> <UnstableFeatures as StyleEditionDefault>::ConfigType

pub fn disable_all_formatting( &self, ) -> <DisableAllFormatting as StyleEditionDefault>::ConfigType

pub fn skip_children(&self) -> <SkipChildren as StyleEditionDefault>::ConfigType

pub fn hide_parse_errors( &self, ) -> <HideParseErrors as StyleEditionDefault>::ConfigType

pub fn show_parse_errors( &self, ) -> <ShowParseErrors as StyleEditionDefault>::ConfigType

pub fn error_on_line_overflow( &self, ) -> <ErrorOnLineOverflow as StyleEditionDefault>::ConfigType

pub fn error_on_unformatted( &self, ) -> <ErrorOnUnformatted as StyleEditionDefault>::ConfigType

pub fn ignore(&self) -> <Ignore as StyleEditionDefault>::ConfigType

pub fn verbose(&self) -> <Verbose as StyleEditionDefault>::ConfigType

pub fn file_lines(&self) -> <FileLinesConfig as StyleEditionDefault>::ConfigType

pub fn emit_mode(&self) -> <EmitModeConfig as StyleEditionDefault>::ConfigType

pub fn make_backup(&self) -> <MakeBackup as StyleEditionDefault>::ConfigType

pub fn print_misformatted_file_names( &self, ) -> <PrintMisformattedFileNames as StyleEditionDefault>::ConfigType

pub(crate) fn default_with_style_edition(style_edition: StyleEdition) -> Config

pub fn set(&mut self) -> ConfigSetter<'_>

pub fn set_cli(&mut self) -> CliConfigSetter<'_>

pub fn was_set(&self) -> ConfigWasSet<'_>

pub fn was_set_cli(&self) -> CliConfigWasSet<'_>

fn fill_from_parsed_config(self, parsed: PartialConfig, dir: &Path) -> Config

pub(crate) fn is_valid_name(name: &str) -> bool

pub fn is_valid_key_val(key: &str, val: &str) -> bool

pub fn used_options(&self) -> PartialConfig

pub fn all_options(&self) -> PartialConfig

pub fn override_value(&mut self, key: &str, val: &str)

pub fn is_hidden_option(name: &str) -> bool

pub fn print_docs(out: &mut dyn Write, include_unstable: bool)

fn set_width_heuristics(&mut self, heuristics: WidthHeuristics)

fn set_heuristics(&mut self)

fn set_ignore(&mut self, dir: &Path)

fn set_merge_imports(&mut self)

fn set_fn_args_layout(&mut self)

fn set_hide_parse_errors(&mut self)

fn set_version(&mut self)

pub fn is_default(&self, key: &str) -> bool

Returns true if the config key was explicitly set and is the default value.

impl Config

pub fn default_for_possible_style_edition( style_edition: Option<StyleEdition>, edition: Option<Edition>, version: Option<Version>, ) -> Config

pub(crate) fn version_meets_requirement(&self) -> bool

pub(crate) fn from_toml_path( file_path: &Path, edition: Option<Edition>, style_edition: Option<StyleEdition>, version: Option<Version>, ) -> Result<Config, Error>

Constructs a Config from the toml file specified at file_path.

This method only looks at the provided path, for a method that searches parents for a rustfmt.toml see from_resolved_toml_path.

Returns a Config if the config could be read and parsed from the file, otherwise errors.

pub(crate) fn from_resolved_toml_path( dir: &Path, edition: Option<Edition>, style_edition: Option<StyleEdition>, version: Option<Version>, ) -> Result<(Config, Option<PathBuf>), Error>

Resolves the config for input in dir.

Searches for rustfmt.toml beginning with dir, and recursively checking parents of dir if no config file is found. If no config file exists in dir or in any parent, a default Config will be returned (and the returned path will be empty).

Returns the Config to use, and the path of the project file if there was one.

pub(crate) fn from_toml(toml: &str, file_path: &Path) -> Result<Config, String>

pub(crate) fn from_toml_for_style_edition( toml: &str, file_path: &Path, edition: Option<Edition>, style_edition: Option<StyleEdition>, version: Option<Version>, ) -> Result<Config, String>