rustfmt_nightly::formatting
Module newline_style

Enums

EffectiveNewlineStyle 🔒

Constants

CARRIAGE_RETURN 🔒

LINE_FEED 🔒

UNIX_NEWLINE 🔒

WINDOWS_NEWLINE 🔒

Functions

apply_newline_style 🔒

Apply this newline style to the formatted text. When the style is set to Auto, the raw_input_text is used to detect the existing line endings.

auto_detect_newline_style 🔒

convert_to_unix_newlines 🔒

convert_to_windows_newlines 🔒

effective_newline_style 🔒

native_newline_style 🔒