Module rustfmt_nightly::macros

Structs

Macro ğŸ”’

MacroArgParser ğŸ”’

Parses macro arguments on macro def.

MacroBranch ğŸ”’

MacroParser ğŸ”’

ParsedMacroArg ğŸ”’

Enums

MacroArg ğŸ”’

MacroArgKind ğŸ”’

MacroPosition ğŸ”’

SpaceState ğŸ”’

Constants

FORCED_BRACKET_MACROS ğŸ”’

Functions

convert_try_mac ğŸ”’

Tries to convert a macro use into a short hand try expression. Returns None when the macro is not an instance of try! (or parsing the inner expression failed).

delim_token_to_str ğŸ”’

force_space_before ğŸ”’

format_lazy_static ğŸ”’

Format lazy_static! from https://crates.io/crates/lazy_static.

format_macro_args ğŸ”’

handle_vec_semi ğŸ”’

ident_like ğŸ”’

last_tok ğŸ”’

macro_style ğŸ”’

next_space ğŸ”’

register_metavariable ğŸ”’

replace_names ğŸ”’

return_macro_parse_failure_fallback ğŸ”’

rewrite_empty_macro_def_body ğŸ”’

rewrite_macro ğŸ”’

rewrite_macro_def ğŸ”’

rewrite_macro_inner ğŸ”’

rewrite_macro_name ğŸ”’

Rewrite macro name without using pretty-printer if possible.

rewrite_macro_with_items ğŸ”’

span_for_token_stream ğŸ”’

wrap_macro_args ğŸ”’

wrap_macro_args_inner ğŸ”’