rustfmt_nightly
Module overflow

Rewrite a list some items with overflow.

Macros

impl_into_overflowable_item_for_ast_node 🔒

impl_into_overflowable_item_for_rustfmt_types 🔒

Structs

Context 🔒

Enums

OverflowableItem 🔒

Constants

SPECIAL_CASE_ATTR 🔒

SPECIAL_CASE_MACROS 🔒

A list of format!-like macros, that take a long format string and a list of arguments to format.

SPECIAL_CASE_MACROS_V2 🔒

Additional special case macros for version 2; these are separated to avoid breaking changes in version 1.

Traits

IntoOverflowableItem 🔒

Functions

can_be_overflowed 🔒

into_overflowable_list 🔒

last_item_shape 🔒

Returns a shape for the last argument which is going to be overflowed.

maybe_get_args_offset 🔒

In case special-case style is required, returns an offset from which we start horizontal layout.

need_block_indent 🔒

no_long_items 🔒

rewrite_with_angle_brackets 🔒

rewrite_with_parens 🔒

rewrite_with_square_brackets 🔒

shape_from_indent_style 🔒