Struct rustfmt_nightly::parse::session::ParseSess

pub(crate) struct ParseSess {
    parse_sess: ParseSess,
    ignore_path_set: Lrc<IgnorePathSet>,
    can_reset_errors: Lrc<AtomicBool>,
}

ParseSess holds structs necessary for constructing a parser.

Fields
parse_sess: ParseSessignore_path_set: Lrc<IgnorePathSet>can_reset_errors: Lrc<AtomicBool>
Implementations

impl ParseSess

pub(crate) fn new(config: &Config) -> Result<ParseSess, ErrorKind>

pub(crate) fn default_submod_path( &self, id: Ident, relative: Option<Ident>, dir_path: &Path ) -> Result<ModulePathSuccess, ModError<'_>>

Determine the submodule path for the given module identifier.

id - The name of the module

relative - If Some(symbol), the symbol name is a directory relative to the dir_path. If relative is Some, resolve the submodle at {dir_path}/{symbol}/{id}.rs or {dir_path}/{symbol}/{id}/mod.rs. if None, resolve the module at {dir_path}/{id}.rs.

dir_path - Module resolution will occur relative to this directory.

pub(crate) fn is_file_parsed(&self, path: &Path) -> bool

pub(crate) fn ignore_file(&self, path: &FileName) -> bool

pub(crate) fn set_silent_emitter(&mut self)

pub(crate) fn span_to_filename(&self, span: Span) -> FileName

pub(crate) fn span_to_file_contents(&self, span: Span) -> Lrc<SourceFile>

pub(crate) fn span_to_first_line_string(&self, span: Span) -> String

pub(crate) fn line_of_byte_pos(&self, pos: BytePos) -> usize

pub(crate) fn byte_pos_same_line(&self, a: BytePos, b: BytePos) -> bool

Determines whether two byte positions are in the same source line.

pub(crate) fn span_to_debug_info(&self, span: Span) -> String

pub(crate) fn inner(&self) -> &RawParseSess

pub(crate) fn snippet_provider(&self, span: Span) -> SnippetProvider

pub(crate) fn get_original_snippet( &self, file_name: &FileName ) -> Option<Lrc<String>>

impl ParseSess

pub(super) fn emit_diagnostics(&self, diagnostics: Vec<Diagnostic>)

pub(super) fn can_reset_errors(&self) -> bool

pub(super) fn has_errors(&self) -> bool

pub(super) fn reset_errors(&self)