Module rustfmt_nightly::utils

Macros

out_of_file_lines_range ğŸ”’

skip_out_of_file_lines_range ğŸ”’

skip_out_of_file_lines_range_visitor ğŸ”’

source ğŸ”’

Traits

NodeIdExt ğŸ”’

Functions

colon_spaces ğŸ”’

contains_skip ğŸ”’

count_lf_crlf ğŸ”’

Returns the number of LF and CRLF respectively.

count_newlines ğŸ”’

depr_skip_annotation ğŸ”’

extra_offset ğŸ”’

filter_attributes ğŸ”’

filtered_str_fits ğŸ”’

first_line_contains_single_line_comment ğŸ”’

first_line_ends_with ğŸ”’

first_line_width ğŸ”’

The width of the first line in s.

format_async ğŸ”’

format_auto ğŸ”’

format_constness ğŸ”’

format_constness_right ğŸ”’

format_defaultness ğŸ”’

format_extern ğŸ”’

format_mutability ğŸ”’

format_unsafety ğŸ”’

format_visibility ğŸ”’

get_prefix_space_width ğŸ”’

indent_next_line ğŸ”’

Based on the given line, determine if the next line can be indented or not. This allows to preserve the indentation of multi-line literals when re-inserted a code block that has been formatted separately from the rest of the code, such as code in macro defs or code blocks doc comments.

inner_attributes ğŸ”’

is_attributes_extendable ğŸ”’

is_block_expr ğŸ”’

is_empty_line ğŸ”’

is_same_visibility ğŸ”’

is_single_line ğŸ”’

is_skip ğŸ”’

is_skip_nested ğŸ”’

last_line_contains_single_line_comment ğŸ”’

last_line_extendable ğŸ”’

last_line_used_width ğŸ”’

The total used width of the last line.

last_line_width ğŸ”’

The width of the last line in s.

left_most_sub_expr ğŸ”’

mk_sp ğŸ”’

mk_sp_lo_plus_one ğŸ”’

outer_attributes ğŸ”’

ptr_vec_to_ref_vec ğŸ”’

remove_trailing_white_spaces ğŸ”’

Removes trailing spaces from the specified snippet. We do not remove spaces inside strings or comments.

rewrite_ident ğŸ”’

semicolon_for_expr ğŸ”’

semicolon_for_stmt ğŸ”’

skip_annotation ğŸ”’

starts_with_newline ğŸ”’

stmt_expr ğŸ”’

trim_left_preserve_layout ğŸ”’

Indent each line according to the specified indent. e.g.

trimmed_last_line_width ğŸ”’

unicode_str_width ğŸ”’

wrap_str ğŸ”’